requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-crispy-forms==1.14.0
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
strgen==1.3.1
